requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
filelock==3.13.1
Pillow==10.1.0
platformdirs==3.11.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6